style: refactor ui.screen, ImplementationNoteRepository and UserCreate
