+++ app/src/androidTest/java/com/github/onlynotesswent/ui/overview/OverviewTest.kt
@@ -13,7 +13,6 @@ import com.github.onlynotesswent.model.users.UserRepository
 import com.github.onlynotesswent.model.users.UserViewModel
 import com.github.onlynotesswent.ui.navigation.NavigationActions
 import com.github.onlynotesswent.ui.navigation.Screen
-import com.github.onlynotesswent.ui.screen.OverviewScreen
 import com.google.firebase.Timestamp
 import org.junit.Before
 import org.junit.Rule
