+++ app/src/main/java/com/github/onlynotesswent/MainActivity.kt
@@ -18,11 +18,11 @@ import com.github.onlynotesswent.ui.authentication.SignInScreen
 import com.github.onlynotesswent.ui.navigation.NavigationActions
 import com.github.onlynotesswent.ui.navigation.Route
 import com.github.onlynotesswent.ui.navigation.Screen
-import com.github.onlynotesswent.ui.screen.AddNoteScreen
-import com.github.onlynotesswent.ui.screen.EditNoteScreen
-import com.github.onlynotesswent.ui.screen.OverviewScreen
+import com.github.onlynotesswent.ui.overview.AddNoteScreen
+import com.github.onlynotesswent.ui.overview.EditNoteScreen
+import com.github.onlynotesswent.ui.overview.OverviewScreen
 import com.github.onlynotesswent.ui.theme.SampleAppTheme
-import com.github.onlynotesswent.ui.user.UserCreate
+import com.github.onlynotesswent.ui.user.CreateUserScreen
 
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
@@ -44,7 +44,7 @@ fun OnlyNotesApp() {
         route = Route.AUTH,
     ) {
       composable(Screen.AUTH) { SignInScreen(navigationActions, userViewModel) }
-      composable(Screen.CREATE_USER) { UserCreate(navigationActions, userViewModel) }
+      composable(Screen.CREATE_USER) { CreateUserScreen(navigationActions, userViewModel) }
     }
 
     navigation(
